% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/map_generator.r
\name{cityChart}
\alias{cityChart}
\title{A control group chart generator function}
\usage{
cityChart(iata_code, exposed_country, number, minimum, maximum)
}
\arguments{
\item{iata_code,}{exposed_country, number, minimum, maximum}
}
\description{
This function charts the cities.
}
\examples{
cityChart()
}
\keyword{chart}

