% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/map_generator.r
\name{cityMap}
\alias{cityMap}
\title{A control group map generator function}
\usage{
cityMap(iata_code, exposed_country, number, minimum, maximum, width = 1600,
  height = 800)
}
\arguments{
\item{iata_code,}{exposed_country, number, minimum, maximum}
}
\description{
This function plots the cities and DMAs on a map.
}
\examples{
cityMap()
}
\keyword{map}

