% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/map_generator.r
\name{getCityList}
\alias{getCityList}
\title{A control group randomizer function}
\usage{
getCityList(iata_code, exposed_country, number, minimum, maximum)
}
\arguments{
\item{iata_code,}{exposed_country, number, minimum, maximum}
}
\description{
This function builds a randomized list of cities.
}
\examples{
getCityList()
}
\keyword{map}

