% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/map_generator.r
\name{inlineChart}
\alias{inlineChart}
\title{A function to put the chart in HTML}
\usage{
inlineChart(iata_code, exposed_country, number, minimum, maximum)
}
\arguments{
\item{iata_code,}{exposed_country, number, minimum, maximum}
}
\description{
This function puts the chart in HTML.
}
\examples{
inlineChart()
}
\keyword{chart}

