% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/map_generator.r
\name{saveChart}
\alias{saveChart}
\title{A chart saving function}
\usage{
saveChart(iata_code, exposed_country, number, minimum, maximum)
}
\arguments{
\item{iata_code,}{exposed_country, number, minimum, maximum}
}
\description{
This function saves the chart.
}
\examples{
saveChart()
}
\keyword{chart}

